+++ src/Integration/Aws/SimpleS3/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- Added optional `versionId` parameter to `SimpleS3Client::getPresignedUrl()`
+
 ## 2.0.0
 
 ### BC-BREAK